float64をやめて全部intに

+++ cmd/test1/bt_label.go
@@ -37,7 +37,7 @@ func (self *BTLabel) Setup(type_ string, data script.ValueMap) error {
 	return nil
 }
 
-func (self *BTLabel) Draw(x, y float64, ctx yamlui.DrawContext) {
+func (self *BTLabel) Draw(x, y int, ctx yamlui.DrawContext) {
 	if x < 0 || y < 0 {
 		return // 負の座標は描画しない
 	}
